python/requirements.txt
# =============================================================================
#  Financial Intelligence Platform — Python Dependencies
#  Version 1.0 · 2026 · HU GAAP
#
#  Installation:
#    pip install -r requirements.txt
#
#  Python version required: 3.11+
#
#  Modules covered:
#    commentary/commentary_generator.py  — AI commentary via Azure OpenAI
#    anomaly_detection/anomaly_detector.py — statistical anomaly detection
#    rag/financial_qa_agent.py           — RAG-based financial Q&A agent
#    forecasting/financial_forecaster.py — Prophet/LSTM revenue & cost forecast
# =============================================================================

# ── Azure SDK ─────────────────────────────────────────────────────────────────
azure-identity==1.17.0               # Managed Identity + Service Principal auth
azure-keyvault-secrets==4.8.0        # Key Vault secret retrieval
azure-storage-blob==12.20.0          # ADLS Gen2 / Blob Storage access
azure-storage-file-datalake==12.15.0 # ADLS Gen2 filesystem operations

# ── Azure OpenAI / AI ─────────────────────────────────────────────────────────
openai==1.30.0                        # Azure OpenAI client (GPT-4o, embeddings)
tiktoken==0.7.0                       # Token counting for prompt size management

# ── Database / Data Access ────────────────────────────────────────────────────
pyodbc==5.1.0                         # ODBC driver for Synapse / Azure SQL
sqlalchemy==2.0.30                    # ORM and connection pooling
pandas==2.2.0                         # DataFrame operations for all scripts
numpy==1.26.0                         # Numerical operations, z-score, etc.

# ── Vector Store / RAG ────────────────────────────────────────────────────────
chromadb>=0.5.0                       # Local vector store for RAG agent (dev/CI)
langchain>=0.2.0                      # LangChain orchestration for RAG pipeline
langchain-openai>=0.1.0               # LangChain Azure OpenAI integration
langchain-community>=0.2.0            # Community document loaders and retrievers

# ── Forecasting ─────────────────────────────────────────────────────
prophet==1.1.5                        # Meta Prophet for monthly revenue/cost forecast
scikit-learn==1.5.0                   # LSTM preprocessing, feature engineering
# tensorflow>=2.16.0                  # LSTM model (optional — heavy)

# ── Statistical Analysis ──────────────────────────────────────────────────────
scipy==1.13.0                         # Z-score, IQR, statistical tests
statsmodels==0.14.0                   # ARIMA, seasonal decomposition, Granger tests

# ── HTTP / API Clients ────────────────────────────────────────────────────────
httpx>=0.27.0                         # Async HTTP client for MNB FX rate API
requests>=2.32.0                      # Sync HTTP client (fallback / simple calls)
lxml>=5.2.0                           # XML parsing for XBRL / Kulcs-Soft exports

# ── Web Frameworks & Auth ──────────────────────────────────────────────────────
fastapi>=0.104.0                      # Fast web framework for APIs
uvicorn>=0.24.0                       # ASGI server for FastAPI
pyjwt>=2.8.0                          # JWT token parsing and validation

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.0                  # .env file loading for local development
pydantic==2.7.0                       # Data validation / settings management
structlog==24.1.0                     # Structured JSON logging for Azure Monitor
tenacity>=8.3.0                       # Retry logic for API calls
schedule==1.2.0                       # In-process job scheduling (dev only)

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
